working on spell choice UI for NewGameView
this implied doing some refactor works to handle enum_simple_map better
for school and spell rarities

+++ src/game/spells/Spells.h
@@ -13,7 +13,7 @@ class Skill;
 
 
 
-enum class SpellRarity : u8
+enum class SpellRarity : u32
 {
   COMMON = 0,
   UNCOMMON,
